rename uint256 to uint

+++ src/Swap.sol
@@ -13,9 +13,9 @@ import "./Socks.sol";
 /// The property above must be true for the swap market itself, regardless of calls to `inflate`.
 contract SocksSwap is ERC1155TokenReceiver {
     Socks constant socks = Socks(0xdfcCFA821F0bFD9d90746021094FAf6C0f10AB63);
-    uint256 constant BOOTSTRAP = 10;
+    uint constant BOOTSTRAP = 10;
 
-    mapping(address => mapping(Sock => uint256)) balanceOf;
+    mapping(address => mapping(Sock => uint)) balanceOf;
 
     /// @notice Bootstraps the pool.
     constructor() {
@@ -26,9 +26,9 @@ contract SocksSwap is ERC1155TokenReceiver {
     /// @notice Mint more socks for the pool contract. Will fail from 9.11.2022.
     function inflate(Sock expected) public {
         // Make sure we fill up with a specific needed side.
-        require(block.difficulty % 2 == uint256(expected));
+        require(block.difficulty % 2 == uint(expected));
 
-        for (uint256 i = 0; i < BOOTSTRAP;) {
+        for (uint i = 0; i < BOOTSTRAP;) {
             socks.mint();
             unchecked {
                 ++i;
@@ -38,13 +38,13 @@ contract SocksSwap is ERC1155TokenReceiver {
 
     /// @notice Optimized swap path if the pool has enough tokens.
     function swap(Sock from, Sock to) external {
-        socks.safeTransferFrom(msg.sender, address(this), uint256(from), 1, "");
-        socks.safeTransferFrom(address(this), msg.sender, uint256(to), 1, "");
+        socks.safeTransferFrom(msg.sender, address(this), uint(from), 1, "");
+        socks.safeTransferFrom(address(this), msg.sender, uint(to), 1, "");
     }
 
     /// @notice If there is nothing on the other side of the swap users might need to deposit and swap at different times.
     function deposit(Sock id) external {
-        socks.safeTransferFrom(msg.sender, address(this), uint256(id), 1, "");
+        socks.safeTransferFrom(msg.sender, address(this), uint(id), 1, "");
         unchecked {
             ++balanceOf[msg.sender][id];
         }
@@ -53,22 +53,22 @@ contract SocksSwap is ERC1155TokenReceiver {
     /// @notice If there is nothing on the other side of the swap users might need to deposit and swap at different times.
     function swapAfterDeposit(Sock from, Sock to) external {
         require(from != to);
-        require(socks.balanceOf(address(this), uint256(to)) > 0);
+        require(socks.balanceOf(address(this), uint(to)) > 0);
 
         --balanceOf[msg.sender][from];
-        socks.safeTransferFrom(address(this), msg.sender, uint256(to), 1, "");
+        socks.safeTransferFrom(address(this), msg.sender, uint(to), 1, "");
     }
 
     /// @notice Users can always withdraw their tokens if a swap was not available for a while.
     function withdraw() external {
-        (uint256 balLeft, uint256 balRight) = (balanceOf[msg.sender][Sock.Left], balanceOf[msg.sender][Sock.Right]);
+        (uint balLeft, uint balRight) = (balanceOf[msg.sender][Sock.Left], balanceOf[msg.sender][Sock.Right]);
         balanceOf[msg.sender][Sock.Left] = 0;
         balanceOf[msg.sender][Sock.Right] = 0;
         if (balLeft > 0) {
-            socks.safeTransferFrom(address(this), msg.sender, uint256(Sock.Left), balLeft, "");
+            socks.safeTransferFrom(address(this), msg.sender, uint(Sock.Left), balLeft, "");
         }
         if (balRight > 0) {
-            socks.safeTransferFrom(address(this), msg.sender, uint256(Sock.Right), balRight, "");
+            socks.safeTransferFrom(address(this), msg.sender, uint(Sock.Right), balRight, "");
         }
     }
 }
